Added componenents to split the message list part.
